navigation bar commmit. Added Lanas code to mine and changed navbar file to Navbuttons as there is already a Navbar component
